+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed
+
+* Position updates can include `Side::UNDEFINED` when reporting
+  net positions.
+
 ## 0.4.5 &ndash; 2020-11-09
 
 ### Changed